glTF working very well so far, deleted alternative header
